Graduate Docusaurus template in C3
