+++ .changeset/tasty-apes-walk.md
@@ -1,0 +1,5 @@
+---
+"create-cloudflare": patch
+---
+
+feat: graduate Docusaurus Workers + Assets template from experimental
