+++ packages/create-cloudflare/src/templates.ts
@@ -20,7 +20,6 @@ import {
 } from "helpers/files";
 import angularTemplateExperimental from "templates-experimental/angular/c3";
 import astroTemplateExperimental from "templates-experimental/astro/c3";
-import docusaurusTemplateExperimental from "templates-experimental/docusaurus/c3";
 import gatsbyTemplateExperimental from "templates-experimental/gatsby/c3";
 import assetsOnlyTemplateExperimental from "templates-experimental/hello-world-assets-only/c3";
 import helloWorldWithDurableObjectAssetsTemplateExperimental from "templates-experimental/hello-world-durable-object-with-assets/c3";
@@ -184,7 +183,6 @@ export function getFrameworkMap({ experimental = false }): TemplateMap {
 		return {
 			angular: angularTemplateExperimental,
 			astro: astroTemplateExperimental,
-			docusaurus: docusaurusTemplateExperimental,
 			gatsby: gatsbyTemplateExperimental,
 			hono: honoTemplateExperimental,
 			next: nextTemplateExperimental,
