+++ packages/create-cloudflare/templates/docusaurus/c3.ts
@@ -1,29 +1,9 @@
-import { runFrameworkGenerator } from "frameworks/index";
-import { detectPackageManager } from "helpers/packageManagers";
-import type { TemplateConfig } from "../../src/templates";
-import type { C3Context } from "types";
+import pages from "./pages/c3";
+import workers from "./workers/c3";
+import type { MultiPlatformTemplateConfig } from "../../src/templates";
 
-const { npm } = detectPackageManager();
-
-const generate = async (ctx: C3Context) => {
-	await runFrameworkGenerator(ctx, [ctx.project.name, "classic"]);
-};
-
-const config: TemplateConfig = {
-	configVersion: 1,
-	id: "docusaurus",
-	frameworkCli: "create-docusaurus",
-	platform: "pages",
+const config: MultiPlatformTemplateConfig = {
 	displayName: "Docusaurus",
-	generate,
-	transformPackageJson: async () => ({
-		scripts: {
-			preview: `${npm} run build && wrangler pages dev ./build`,
-			deploy: `${npm} run build && wrangler pages deploy ./build`,
-		},
-	}),
-	devScript: "preview",
-	deployScript: "deploy",
-	previewScript: "preview",
+	platformVariants: { pages, workers },
 };
 export default config;
