+++ packages/create-cloudflare/templates/docusaurus/workers/c3.ts
@@ -1,6 +1,6 @@
 import { runFrameworkGenerator } from "frameworks/index";
 import { detectPackageManager } from "helpers/packageManagers";
-import type { TemplateConfig } from "../../src/templates";
+import type { TemplateConfig } from "../../../src/templates";
 import type { C3Context } from "types";
 
 const { npm } = detectPackageManager();
@@ -18,7 +18,7 @@ const config: TemplateConfig = {
 	copyFiles: {
 		path: "./templates",
 	},
-	path: "templates-experimental/docusaurus",
+	path: "templates/docusaurus/workers",
 	generate,
 	transformPackageJson: async () => ({
 		scripts: {
