+++ packages/create-cloudflare/templates/docusaurus/workers/templates/wrangler.jsonc
